MAINT: Clean for crackfortran to use capi_maps
